fixed int -> float32 in payload structs

+++ payload.go
@@ -2,20 +2,20 @@ package main
 
 // Subscriber holds info about a subscriber to currency-info in the system
 type Subscriber struct {
-	WebhookURL      string `json:"webhookURL"`
-	BaseCurrency    string `json:"baseCurrency"`
-	TargetCurrency  string `json:"targetCurrency"`
-	MinTriggerValue int    `json:"minTriggerValue"`
-	MaxTriggerValue int    `json:"maxTriggerValue"`
+	WebhookURL      string  `json:"webhookURL"`
+	BaseCurrency    string  `json:"baseCurrency"`
+	TargetCurrency  string  `json:"targetCurrency"`
+	MinTriggerValue float32 `json:"minTriggerValue"`
+	MaxTriggerValue float32 `json:"maxTriggerValue"`
 }
 
 // CurrencyPayload holds info sent to subscriber
 type CurrencyPayload struct {
-	BaseCurrency    string `json:"baseCurrency"`
-	TargetCurrency  string `json:"targetCurrency"`
-	CurrentRate     int    `json:"currentRate"`
-	MinTriggerValue int    `json:"minTriggerValue"`
-	MaxTriggerValue int    `json:"maxTriggerValue"`
+	BaseCurrency    string  `json:"baseCurrency"`
+	TargetCurrency  string  `json:"targetCurrency"`
+	CurrentRate     float32 `json:"currentRate"`
+	MinTriggerValue float32 `json:"minTriggerValue"`
+	MaxTriggerValue float32 `json:"maxTriggerValue"`
 }
 
 // CurrencyRequest holds info received from the user during basic currency
